1/4 problem removed

+++ simulator/exec.h
@@ -3,4 +3,4 @@
 #include<string.h>
 #include<unistd.h>
 #include<math.h>
-void exec(unsigned int*,float*,int,unsigned int*,unsigned int*,int*,FILE*,FILE*,int*,unsigned int*);
+void exec(unsigned int*,float*,int,unsigned int*,unsigned int*,int*,FILE*,FILE*,long long int*,long long int*);
